% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ramlegacy.R
\name{show_sheets}
\alias{show_sheets}
\title{show_sheets}
\usage{
show_sheets(path)
}
\arguments{
\item{path}{Path to the zipped version of the excel database}
}
\value{
A character vector containing the sheet names
}
\description{
Unzips the excel database present at path and displays the sheets present in it
}
